requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
transformers==4.35.2
torch==2.1.1
pydantic==2.5.2
python-dotenv==1.0.0
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
aiohttp==3.9.1
gTTS==2.4.0
pytest==7.4.3
httpx==0.25.2 

